requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
requests==2.32.3
PyYAML==6.0.2
